+++ api/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "nodemon index.js",
-    "dev": "node index.js"
+    "dev": "node src/index.js"
   },
   "author": "",
   "license": "ISC",
