+++ api/src/dbconfig/dbconfig.js
@@ -1,0 +1,15 @@
+module.exports = {
+host: 'db',
+user: 'root',
+password: 'pass',
+database: "mydb"
+}
+    
+  //   dialect: 'mysql',
+  // host: 'db',
+  // username: 'root',
+  // password: 'pass',
+  // database: "mydb",
+  // define:{
+  // timestamps: true,
+  // underscored: true, }
