+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "body-parser": "^1.19.0"
+  }
+}
